requirement-freeze.txt
annotated-types==0.7.0
anyio==4.6.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.31.0
blinker==1.8.2
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cymem==2.0.8
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
fuzzywuzzy==0.18.0
blis==1.0.1
numpy>=2.0.0,<3.0.0
gensim==4.1.2  # or a version that works with numpy >=2.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
Levenshtein==0.26.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msrest==0.7.1
murmurhash==1.0.10
networkx==3.3
nltk==3.9.1
oauthlib==3.2.2
openai==1.51.0
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
preshed==3.0.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMuPDF==1.24.11
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-Levenshtein==0.26.0
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
safetensors==0.4.5
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.3
textwrap3==0.9.2
thinc==8.3.2
tokenizers==0.20.0
torch==2.2.2
tqdm==4.66.5
transformers==4.45.1
typer==0.12.5
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0
